moves oidc header set logic from route to controller
